Fix agents and commits, tests are green

+++ lib/src/datetime_helpers.rs
@@ -1,7 +1,7 @@
 /// Returns the current timestamp
-pub fn now() -> u128 {
+pub fn now() -> u64 {
   std::time::SystemTime::now()
     .duration_since(std::time::UNIX_EPOCH)
     .expect("You're a time traveler")
-    .as_millis()
+    .as_millis() as u64
 }
